+++ src/layout.php
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Signature guide - Fintecture</title>
     <link rel="icon" href="assets/favicon.png">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" integrity="sha256-MBffSnbbXwHCuZtgPYiwMQbfE7z+GOZ7fBPCNB06Z98=" crossorigin="anonymous">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css" integrity="sha256-9kPW/n5nn53j4WMRYAxe9c1rCY96Oogo/MKSVdKzPmI=" crossorigin="anonymous">
     <link rel="stylesheet" href="assets/style.css">
     <link rel="stylesheet" href="assets/prism.css">
@@ -25,7 +25,10 @@
                         <a href="https://docs.fintecture.com" class="nav-link"><i class="bi bi-box-arrow-up-right"></i> API Documentation</a>
                     </li>                    
                     <li class="nav-item">
-                        <a href="https://github.com/Fintecture/signature-guide/tree/src/scripts" class="nav-link"><i class="bi bi-file-earmark-code"></i> Complete examples</a>
+                        <a href="https://github.com/Fintecture/signature-guide/tree/main/src/scripts" class="nav-link"><i class="bi bi-file-earmark-code"></i> Complete examples</a>
+                    </li>
+                    <li class="nav-item">
+                        <a href="https://github.com/Fintecture/signature-guide/tree/main/src/scripts" class="nav-link"><i class="bi bi-github"></i> Source code</a>
                     </li>
                 </ul>
             </header>
@@ -38,7 +41,7 @@
         </div>
     </main>
 
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.min.js" integrity="sha256-YMa+wAM6QkVyz999odX7lPRxkoYAan8suedu4k2Zur8=" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
     <script src="assets/prism.js"></script>
     <script src="assets/main.js"></script>
 </body>