+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## 2023-02-07
+## 2024-03-11
+
+- Publish repo on GitHub
+- Complete scripts examples
+
+## 2024-02-07
 
 - Add JS steps to signature guide